Remove indirect dependency on micro_rpc from restricted kernel
extracting the core oak channel logic to a lightweight core crate
