+++ oak_channel_core/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "oak_channel_core"
+version = "0.1.0"
+edition = "2021"
+license = "Apache-2.0"
+
+[features]
+default = []
+std = []
+
+[dependencies]
+anyhow = { version = "*", default-features = false }
