+++ oak_restricted_kernel/src/lib.rs
@@ -73,7 +73,7 @@ use mm::{
     frame_allocator::PhysicalMemoryAllocator, page_tables::RootPageTable,
     virtual_address_allocator::VirtualAddressAllocator,
 };
-use oak_channel::Channel;
+use oak_channel_core::Channel;
 use oak_core::sync::OnceCell;
 use oak_linux_boot_params::{BootParams, Ramdisk};
 use oak_sev_guest::msr::{change_snp_state_for_frame, get_sev_status, PageAssignment, SevStatus};
@@ -387,8 +387,11 @@ pub fn start_kernel(info: &BootParams) -> ! {
                 // We need to load the application binary before we hand the channel over to the
                 // syscalls, which expose it to the user space.
                 info!("Loading application binary...");
-                payload::Application::load_raw(&mut *channel)
-                    .expect("failed to load application binary from channel")
+                let payload = oak_channel_core::basic_framed::load_raw(&mut *channel)
+                    .expect("failed to load application binary from channel");
+                log::info!("Binary loaded, size: {}", payload.len());
+                payload::Application::new(payload.into_boxed_slice())
+                    .expect("failed to parse application from ramdisk")
             }
         }
     };
